Move tests to CollectionExpressionTests
