+++ src/Compilers/CSharp/Test/EndToEnd/EndToEndTests.cs
@@ -721,83 +721,5 @@ $@"        if (F({i}))
             Assert.Collection(runResult.TrackedSteps["result_ForAttributeWithMetadataName"],
                 step => Assert.True(step.Outputs.Single().Value is ClassDeclarationSyntax { Identifier.ValueText: "C1" }));
         }
-
-        [Fact, WorkItem("https://github.com/dotnet/roslyn/issues/70381")]
-        public void ExtremelyNestedCollectionExpressionDoesNotOverflow_1()
-        {
-            var code = $$"""
-                #nullable enable
-                using System;
-                using System.Collections;
-                using System.Collections.Generic;
-
-                class MyCollection : IEnumerable
-                {
-                    void M()
-                    {
-                        MyCollection collection = {{new string('[', 64)}}{{new string(']', 64)}};
-                    }
-
-                    public void Add(MyCollection c)
-                    {
-
-                    }
-
-                    public IEnumerator<int> GetEnumerator() => throw new NotImplementedException();
-
-                    IEnumerator IEnumerable.GetEnumerator() => throw new NotImplementedException();
-                }
-                """;
-
-            var compilation = CreateCompilation(code);
-            CompileAndVerify(compilation);
-
-            var tree = compilation.SyntaxTrees[0];
-            var model = compilation.GetSemanticModel(tree);
-            var collectionExpression = tree.GetRoot().DescendantNodes().OfType<CollectionExpressionSyntax>().First();
-
-            Assert.NotNull(model.GetOperation(collectionExpression));
-        }
-
-        [Fact, WorkItem("https://github.com/dotnet/roslyn/issues/70381")]
-        public void ExtremelyNestedCollectionExpressionDoesNotOverflow_2()
-        {
-            var code = $$"""
-                #nullable enable
-                using System;
-                using System.Collections;
-                using System.Collections.Generic;
-
-                class MyCollection : IEnumerable
-                {
-                    void M()
-                    {
-                        MyCollection collection = {{new string('[', 65)}}{{new string(']', 65)}};
-                    }
-
-                    public void Add(MyCollection c)
-                    {
-
-                    }
-
-                    public IEnumerator<int> GetEnumerator() => throw new NotImplementedException();
-
-                    IEnumerator IEnumerable.GetEnumerator() => throw new NotImplementedException();
-                }
-                """;
-
-            var compilation = CreateCompilation(code);
-            compilation.VerifyDiagnostics(
-                // (10,99): error CS8078: An expression is too long or complex to compile
-                //         MyCollection collection = [[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]];
-                Diagnostic(ErrorCode.ERR_InsufficientStack, "[]").WithLocation(10, 99)
-            );
-
-            var tree = compilation.SyntaxTrees[0];
-            var model = compilation.GetSemanticModel(tree);
-            var collectionExpression = tree.GetRoot().DescendantNodes().OfType<CollectionExpressionSyntax>().First();
-
-            Assert.NotNull(model.GetOperation(collectionExpression));
-        }
     }
 }
